Split tests for classes RHEALPixDGGS and Cell into two separate scripts (following split of the two classes into separate scripts)
